requirements.txt
black==21.12b0
djangorestframework==3.13.1
pip-chill==1.0.1
psycopg2==2.9.3
python-dotenv==0.19.2
whitenoise==5.3.0
